+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,19 @@
 --->
 
 ## Upcoming Release
+* Added new cmdlet for List NetworkSecurityPerimeter ServiceTags
+    - `Get-AzNetworkSecurityPerimeterServiceTag`
+* Added properties 'DedicatedBackendConnection', 'ValidateCertChainAndExpiry', 'ValidateSNI', and 'SniName' to Application Gateway Backend HTTP Settings, as well as support for them in the following cmdlets:
+    - `New-AzApplicationGatewayBackendHttpSetting`
+    - `Add-AzApplicationGatewayBackendHttpSetting`
+    - `Set-AzApplicationGatewayBackendHttpSetting`
+* Onboarded Maps to Private Link Common Cmdlets
+* Added cmdlet `Get-AzVirtualNetworkGatewayFailoverSingleTestDetails` to retrieve details of a single failover test on a virtual network gateway.
+* Added cmdlet `Get-AzVirtualNetworkGatewayFailoverAllTestsDetails` to list all failover tests for a virtual network gateway.
+* Added cmdlet `Start-AzVirtualNetworkGatewaySiteFailoverTest` to initiate a site failover test on a virtual network gateway.
+* Added cmdlet `Stop-AzVirtualNetworkGatewaySiteFailoverTest` to stop an ongoing site failover test.
+* Added cmdlet `Get-AzVirtualNetworkGatewayResiliencyInformation` to retrieve resiliency information for a virtual network gateway.
+* Added cmdlet `Get-AzVirtualNetworkGatewayRoutesInformation` to get route information for a virtual network gateway.
 
 ## Version 7.19.1
 * Added a read-only property `ComputedDisabledRules` to `ApplicationGatewayFirewallPolicyManagedRuleSet`. This property shows which rules are effectively disabled, based on both user-defined WAF policy overrides and the default state of the rules in the managed ruleset.
@@ -31,7 +44,6 @@
         - `New-AzApplicationGatewayFirewallPolicy`
 * Onboarded FileShares to Private Link Common Cmdlets
 * Updated Api version for StorageSync to Private Link Common Cmdlets
-
 
 ## Version 7.19.0
 * Returned appgw and agc in waf policy